verify blobstream address
