+++ crates/proofs/src/lib.rs
@@ -5,3 +5,4 @@
 extern crate alloc;
 
 pub mod blobstream_inclusion;
+pub mod types;
